refactor: SettingsFragment MVP 구조 제거
