+++ app/src/main/res/layout/fragment_settings.xml
@@ -12,7 +12,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".ui.settings.view.SettingsFragment">
+        tools:context=".ui.settings.SettingsFragment">
 
         <TextView
             android:id="@+id/tv_push_notification_title"
